Add new Blog/Newsroom page models
