+++ cfgov/jinja2/v1/_includes/templates/render_block.html
@@ -1,6 +1,6 @@
 
 {% macro render(block, index) %}
-    <div class="block {{ 'block__flush-top' if loop.index == 1 else '' }} {{ block.block.meta.classname if block.block.meta.classname else '' -}}">
+    <div class="block {{ 'block__flush-top' if index == 1 else '' }} {{ block.block.meta.classname if block.block.meta.classname else '' -}}">
         {{ render_stream_child(block) }}
     </div>
 {% endmacro %}
