+++ cfgov/v1/models/__init__.py
@@ -12,6 +12,8 @@ from .browse_filterable_page import *
 from .learn_page import *
 from .home_page import *
 from .sublanding_filterable_page import *
+from .blog_page import *
+from .newsroom_page import *
 
 import os
 from django.conf import settings
